refactor: move event utilities from interfaces/events to models/events
