+++ axis/interfaces/event_instances.py
@@ -9,9 +9,9 @@ from ..models.event_instance import (
     ListEventInstancesRequest,
     ListEventInstancesResponse,
 )
+from ..models.events.topic_normalizer import to_canonical, to_topic_filter
 from .api_handler import ApiHandler
 from .event_manager import BLACK_LISTED_TOPICS
-from .events.topic_normalizer import to_canonical, to_topic_filter
 
 if TYPE_CHECKING:
     from ..models.event import Event
