--- axis/interfaces/events/__init__.py
@@ -1,1 +1,0 @@
-"""Event-related interface helpers and contracts."""
